% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dirOperating.R
\name{dir.getAllFileBaseName}
\alias{dir.getAllFileBaseName}
\title{获取指定目录下的所有文件名}
\usage{
dir.getAllFileBaseName(dir_name)
}
\arguments{
\item{dir_name}{目录}
}
\value{
返回值
}
\description{
获取指定目录下的所有文件名
}
\examples{
dir.getAllFileBaseName("./data-raw");
}
