% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dirOperating.R
\name{dir.getAllFileFullName}
\alias{dir.getAllFileFullName}
\title{获取指定目录下所有文件的全称}
\usage{
dir.getAllFileFullName(dir_name)
}
\arguments{
\item{dir_name}{文件目录}
}
\value{
返回值，此目录下所有文件的全称
}
\description{
获取指定目录下所有文件的全称
}
\examples{
dir.getAllFileFullName('./data-raw/')
}
