% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-fileOperating.R
\name{file.getFullDir}
\alias{file.getFullDir}
\title{根据文件的全名来获取文件的直接上级路径}
\usage{
file.getFullDir(fullFile = "./data-raw/file/test.aa.xlsx")
}
\arguments{
\item{fullFile}{文件全名}
}
\value{
返回值
}
\description{
根据文件的全名来获取文件的直接上级路径
}
\examples{
file.getFullDir('./data-raw/file/test.aa.xlsx');
}
