% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-fileOperating.R
\name{file.postAddName}
\alias{file.postAddName}
\title{添加基础文件名}
\usage{
file.postAddName(baseName, sep = "\\\\.", FlagText = "处理后")
}
\arguments{
\item{baseName}{基础文件名}

\item{sep}{正则表达式}

\item{FlagText}{添加内容}
}
\value{
返回值
}
\description{
添加基础文件名
}
\examples{
 file.postAddName('aaa.xlsx');
}
