% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-fileOperating.R
\name{file.replaceName}
\alias{file.replaceName}
\title{替换文件名称}
\usage{
file.replaceName(fullFileNames, newShortNames)
}
\arguments{
\item{fullFileNames}{旧的文件全名}

\item{newShortNames}{新的短的文件名}
}
\value{
返回值 字符串
}
\description{
替换文件名称
}
\examples{
file.replaceName('./data-raw/file/test/')
}
