% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-fileOperating.R
\name{file.writeCharEachLine}
\alias{file.writeCharEachLine}
\title{将字符向量元素的每一元素写成一行}
\usage{
file.writeCharEachLine(charVector, file = "")
}
\arguments{
\item{charVector}{字符向量}

\item{file}{文件路径}
}
\value{
返回值
}
\description{
将字符向量元素的每一元素写成一行
}
\examples{
file.writeCharEachLine(letters,file="aa.txt");
}
