% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-fileOperating.R
\name{file.writeObject}
\alias{file.writeObject}
\title{将R对象写入文件}
\usage{
file.writeObject(..., file = "", append = TRUE)
}
\arguments{
\item{...}{列表对象,写入内容}

\item{file}{文件名}

\item{append}{是否追加到文件，默认为是}
}
\value{
返回值，空
}
\description{
将R对象写入文件
}
\examples{
file.writeObject(letters,file="aa.txt")
}
